# Conflicts:
#	.idea/caches/build_file_checksums.ser
#	app/src/main/res/layout/notification_items.xml

+++ app/src/main/java/com/nisarga/nisargaveggiez/MyOrder/MyOrders.java
@@ -66,8 +66,6 @@ public class MyOrders extends AppCompatActivity {
             getSupportActionBar().setDisplayHomeAsUpEnabled(true);
             getSupportActionBar().setDisplayShowHomeEnabled(true);
 
-          //  orderItem = new orderItem(MyOrders.this);
-
         apiInterface = APIClient.getClient().create(APIInterface.class);
 
         if (Utils.CheckInternetConnection(getApplicationContext())) {
@@ -83,18 +81,25 @@ public class MyOrders extends AppCompatActivity {
                     List<MyOrderList.MyOrderListDatum> datumList1 = resource.result;
                     if((resource.mstatus).equals("success"))
                     {
-                    for (MyOrderList.MyOrderListDatum orderList : datumList1) {
+                        if(datumList1.size()>0) {
+                            for (MyOrderList.MyOrderListDatum orderList : datumList1) {
 
-                            session.storeStatusOrder(orderList.sstatus);
-                            session.storeCancelId(orderList.scancel);
-                            mCartView.addView(new orderItem(MyOrders.this, orderList.sorder_id, orderList.sdate_added,
-                                    orderList.sstatus, orderList.scancel));
+                                session.storeStatusOrder(orderList.sstatus);
+                                session.storeCancelId(orderList.scancel);
+                                mCartView.addView(new orderItem(MyOrders.this, orderList.sorder_id, orderList.sdate_added,
+                                        orderList.sstatus, orderList.scancel));
 
-                            Log.e("-----OrdersList--", orderList.sfirstname + " " + orderList.scancel);
+                                Log.e("-----OrdersList--", orderList.sfirstname + " " + orderList.scancel);
 
+                            }
+                        }
+                        else
+                        {
+                            Toast.makeText(getApplicationContext(), "No Orders", Toast.LENGTH_LONG).show();
                         }
 
                     }
+
                     else
                     {
                         Toast.makeText(getApplicationContext(),resource.msg,Toast.LENGTH_SHORT).show();
